requirements.txt
allure-pytest==2.13.2
pytest==7.3.2
selenium==4.10.0